+++ front/src/modules/activities/components/TaskRow.tsx
